requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
Django==3.1
django-storages==1.10
google-api-core==1.22.1
google-auth==1.21.0
google-cloud-core==1.4.1
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
idna==2.10
mysqlclient==2.0.1
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
